+++ src/common/utils/computeGroupId/computeGroupId.ts
