+++ src/common/utils/computeGroupId/index.ts
@@ -1,0 +1,1 @@
+export { default } from './computeGroupId';
