+++ src/common/utils/createLogger/index.ts
@@ -1,0 +1,1 @@
+export { default } from './createLogger';
