+++ src/common/utils/formatCurrencyUnit/formatCurrencyUnit.test.ts
