+++ src/extension/components/AssetsTab/AssetTabArc200AssetItem.tsx
@@ -44,7 +44,8 @@ import AccountService from '@extension/services/AccountService';
 import { IAccount, IArc200Asset, INetwork } from '@extension/types';
 
 // utils
-import { convertToStandardUnit, formatCurrencyUnit } from '@common/utils';
+import convertToStandardUnit from '@common/utils/convertToStandardUnit';
+import formatCurrencyUnit from '@common/utils/formatCurrencyUnit';
 
 interface IProps {
   account: IAccount;
