+++ src/extension/components/AssetsTab/AssetTabStandardAssetItem.tsx
@@ -41,7 +41,8 @@ import AccountService from '@extension/services/AccountService';
 import { IAccount, INetwork, IStandardAsset } from '@extension/types';
 
 // utils
-import { convertToStandardUnit, formatCurrencyUnit } from '@common/utils';
+import convertToStandardUnit from '@common/utils/convertToStandardUnit';
+import formatCurrencyUnit from '@common/utils/formatCurrencyUnit';
 
 interface IProps {
   account: IAccount;
