+++ src/extension/features/add-asset/thunks/queryArc200AssetThunk.ts
@@ -30,7 +30,7 @@ import {
 } from '../types';
 
 // utils
-import { getIndexerClient } from '@common/utils';
+import getIndexerClient from '@common/utils/getIndexerClient';
 import {
   convertGenesisHashToHex,
   selectAssetsForNetwork,
