+++ src/extension/features/send-assets/thunks/submitTransactionThunk.ts
@@ -31,11 +31,9 @@ import {
 } from '@extension/types';
 
 // utils
-import {
-  convertToAtomicUnit,
-  getAlgodClient,
-  getIndexerClient,
-} from '@common/utils';
+import convertToAtomicUnit from '@common/utils/convertToAtomicUnit';
+import getAlgodClient from '@common/utils/getAlgodClient';
+import getIndexerClient from '@common/utils/getIndexerClient';
 import { selectNetworkFromSettings } from '@extension/utils';
 import {
   sendArc200AssetTransferTransaction,
