+++ src/extension/hooks/useOnDebugLogging/useOnDebugLogging.ts
@@ -12,7 +12,7 @@ import { ILogger } from '@common/types';
 import { IAppThunkDispatch } from '@extension/types';
 
 // utils
-import { createLogger } from '@common/utils';
+import createLogger from '@common/utils/createLogger';
 
 export default function useOnDebugLogging(): void {
   const dispatch: IAppThunkDispatch = useDispatch<IAppThunkDispatch>();
