+++ src/extension/modals/AddAssetModal/AddAssetModalArc200AssetSummaryContent.tsx
@@ -31,7 +31,8 @@ import {
 } from '@extension/types';
 
 // utils
-import { convertToStandardUnit, formatCurrencyUnit } from '@common/utils';
+import convertToStandardUnit from '@common/utils/convertToStandardUnit';
+import formatCurrencyUnit from '@common/utils/formatCurrencyUnit';
 
 interface IProps {
   asset: IArc200Asset;
