+++ src/extension/modals/AddAssetModal/AddAssetModalStandardAssetSummaryContent.tsx
@@ -42,7 +42,8 @@ import {
 } from '@extension/types';
 
 // utils
-import { convertToStandardUnit, formatCurrencyUnit } from '@common/utils';
+import convertToStandardUnit from '@common/utils/convertToStandardUnit';
+import formatCurrencyUnit from '@common/utils/formatCurrencyUnit';
 import { createIconFromDataUri, isAccountKnown } from '@extension/utils';
 
 interface IProps {
