+++ src/extension/modals/SendAssetModal/SendAmountInput.tsx
@@ -41,7 +41,8 @@ import {
 } from '@extension/types';
 
 // utils
-import { convertToStandardUnit, formatCurrencyUnit } from '@common/utils';
+import convertToStandardUnit from '@common/utils/convertToStandardUnit';
+import formatCurrencyUnit from '@common/utils/formatCurrencyUnit';
 import { convertGenesisHashToHex } from '@extension/utils';
 
 interface IProps {
