+++ src/extension/modals/SignTxnsModal/MultipleTransactionsContent.tsx
@@ -32,7 +32,7 @@ import {
 } from '@extension/types';
 
 // utils
-import { computeGroupId } from '@common/utils';
+import computeGroupId from '@common/utils/computeGroupId';
 import { parseTransactionType } from '@extension/utils';
 
 interface IProps {
