+++ src/extension/modals/SignTxnsModal/PaymentTransactionContent.tsx
@@ -22,7 +22,8 @@ import { IAccount, IAccountInformation, INetwork } from '@extension/types';
 import { ICondensedProps } from './types';
 
 // utils
-import { convertToStandardUnit, formatCurrencyUnit } from '@common/utils';
+import convertToStandardUnit from '@common/utils/convertToStandardUnit';
+import formatCurrencyUnit from '@common/utils/formatCurrencyUnit';
 import { createIconFromDataUri, parseTransactionType } from '@extension/utils';
 import AccountService from '@extension/services/AccountService';
