+++ tsconfig.json
@@ -22,7 +22,7 @@
       "@common/errors": ["src/common/errors"],
       "@common/messages": ["src/common/messages"],
       "@common/types": ["src/common/types"],
-      "@common/utils": ["src/common/utils"],
+      "@common/utils/*": ["src/common/utils/*"],
       "@extension/components/*": ["src/extension/components/*"],
       "@extension/config": ["src/extension/config"],
       "@extension/constants": ["src/extension/constants"],
